Add a test to make sure that the module component serializer is idempotent
This commit introduces a test case which works on sample module metadata files. It will enforce that once metadata has been
parsed and serialized by the module serializer, then when it's read back, metadata contents is the same. Technically speaking
it's not a unit test, nor an integration test. It's a bit in between, as it will ensure that both seralization and deserialization
works, for a set of reasonable inputs.

The idea would be that each time a new feature is introduced in Gradle metadata files, we add a sample covering this case. This
will ensure that the resolvers will work consistently, independently of where metadata comes from (binary cache or parsed metadata).